requirements.txt
opencv-python
mediapipe
python-socketio[client]
flask
eventlet
Pillow
pyautogui
requests